Finetuning stuff
- auto configuration using spring.factories
- h2 console
- removed @Version and @LastModifiedDate from Pizza
- implicit init() for PizzaInitializer
- default banner ;-)
