+++ catalog/src/main/java/com/epages/microservice/handson/catalog/Topping.java
@@ -10,6 +10,5 @@ public enum Topping {
     TUNA,
     NUTELLA,
     PEANUT_BUTTER,
-    VEGEMITE;
-
+    VEGEMITE
 }
