+++ shared/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,5 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+com.epages.microservice.handson.shared.jpa.JpaAutoConfiguration,\
+com.epages.microservice.handson.shared.event.EventAutoConfiguration,\
+com.epages.microservice.handson.shared.json.JsonAutoConfiguration,\
+com.epages.microservice.handson.shared.web.WebAutoConfiguration
