+++ shared/src/main/resources/banner.txt
@@ -1,0 +1,41 @@
+                                    __..---'"""""""`---..__
+                               _.-''    ____......____     ``-._
+                           _.-'   _.--'"     _        `""-.._   `-._
+                        _,'   _.-'    *  _.-"_`   ~  .-.  ,  `-._   `._
+                      ,'    .'      ~    `.-'    "  :   :   ~    `-.   `.
+                    ,'    .'   "      "     ____     ._.      /. '  `.   `.
+                  ,'   .- .--.  *  ||  ,  ."    `.        "  //   "   `.   `.
+                ,'   .'  :   .'    ||     """| |"'  ~   _   // ,    *   `.   `.
+               /   .' ~  : ,'`,  ~ |'   "    '-'  *    (C)      ."".    " `.   \
+              /   '       ` `'         *   ~     .   "     *   :   .  ~     .   \
+             /   '   // *   ,  ,-.  ~   _   ,  .','     '    "  `.'      (0) .   \
+            /   .   //   (c)  :   :    (o)   .'.'   ,  ~  \\         *       *.   \
+           /   . , //  ,      `.__,   "       '      .--.  \\  ~   =====   "   .   \
+          /   '           "        ____  ,  "   *   `.   `  })  _  "   ~  .     .   \
+         .   '   "  ,  ~    _ *    ----'  ~       , .'`, ;     (O)    .   ,-. "  .
+         '  '  _      .    (o)  ~       *   " (D)    `" `    "     *   ___| |___  .  .
+        j  .  (c) *  `.`.            "    .-     .  ~    , *   .       '._    .'  |  |
+        |  | "         `.'  "   //      .'  :  "      '         ~  //     `"""   *|  |
+        |  |       "  .        //   ~  '    ,    ~ , ===="  (c)   //  ~    .    " |  |
+        |  |  ~ ,'    .-. ~      ,  _  `...' (@)        ~  `           *     (C)  |  |
+        |  |*   ||  __| |___  "    (0) "     __   ,""-.      .--.  "  _   '       |  |
+        |  | ,  ||  '.    ,'    ~    __..--"' ,  '   .'   " '    .   ())    \\  ~ |  |
+        |  |      ~   `""" __,..--'"'        ,   '_,'`. *    `--"        . " \\   |  |
+        '  `  "   __...--''          __        *    `.'    ~     "   .--.   , \\  '  |
+         `..`--'''          __..---'' .'    /___..      "   .  *    `.   `    .  .   '
+                  __,..--'''    _  " ,     / ---""  ~    . `.`.  ~  .'`, ;  "   .   .
+         __..---''     ,   "   (O)  ,     /   .       (Q)    `.`.    `" `    * .   ,
+ ,..--'''    .   *  ~   ||   *    ~      / "     *        , ~  `    ~  , (O)  '   .
+|  | ,   (@)   "      " ||   .-.   '    /   _  ~   " || .    *    "          '   .
+'  '   "    ~ .   .--.  |j  :   : /    /~  (O)       ||   "   .  _  *  /. " '   ,
+ .  `    _____   `.   `      .__,/    /         .-.  ||   ,--.  (@)   //  .    /
+  `  `.  ----- * .'`, ;   "     /    /   "     :   :     '   .'      //  .   ,'
+   `.  .  ~       `" `  *   ~  /    /,      ~   .__,  ,  : ,'`  "      .'   /
+     .  `  , (c)     '  .     /    /    *              ~  ' `'    * _.'   ,'
+      `.  .        *  .  (C) /    /  "  (o)  \\ ,  *          ~ _.-"    ,'
+        .  `.   "    // .   /    ,._    ,     \\      (O) * _.-"     _.'
+         `.  `.   , //     /    ,   ``-.._  *  `  "   __.-"'     _.-'
+           `._  `.    . " /     `._       "`--....--""      _,.-'
+              `.  `-._   /         `--..___         ___..--'
+                `-._  ` ,                  `'"""""''
+                    `-.'
